upd types, layouts
