+++ src/app/api/signup/route.ts
@@ -1,6 +1,5 @@
 import dbConnect from "@/lib/dbConnect";
 import User from "@/models/User";
-
 import bcrypt from "bcryptjs";
 import { NextResponse } from "next/server";
