+++ src/features/CheckboxFilter/CheckboxFilter.tsx
@@ -1,8 +1,13 @@
-import { Designer, PriceRange, ProductType, SortCategory } from "@/app/types";
 import CheckBox from "@/components/CheckBox/CheckBox";
 import Dropdown from "@/components/Dropdown/Dropdown";
 import Typography from "@/components/Typography/Typography";
 import useIsMobile from "@/lib/useIsMobile";
+import {
+  Designer,
+  PriceRange,
+  ProductType,
+  SortCategory,
+} from "@/types/product";
 
 import { useRouter } from "next/navigation";
 import React from "react";
