+++ src/lib/products.ts
@@ -1,3 +1,5 @@
+import ProductModel, { IProduct } from "@/models/Product";
+import dbConnect from "./dbConnect";
 import {
   Product,
   ProductCategory,
@@ -5,9 +7,7 @@ import {
   SearchParams,
   SortBy,
   SortOrder,
-} from "@/app/types";
-import ProductModel, { IProduct } from "@/models/Product";
-import dbConnect from "./dbConnect";
+} from "@/types/product";
 
 const transformProduct = (productModel: IProduct): Product => ({
   id_: productModel._id.toString(),
@@ -28,7 +28,7 @@ const transformProduct = (productModel: IProduct): Product => ({
 
 const transformProductListing = (products: IProduct[]): ProductListing[] => {
   return products.map((product) => ({
-    _id: product._id.toString(),
+    id_: product._id.toString(),
     name: product.name,
     price: product.price,
     imageSrc: product.imageSrc,
