+++ src/types/types.ts
@@ -1,0 +1,5 @@
+export interface Link {
+  id: string;
+  title: string;
+  href: string;
+}
